[PM-24468] Cleanup cipher risk module exports
