+++ crates/bitwarden-vault/src/cipher_risk/mod.rs
@@ -1,8 +1,7 @@
-pub(crate) mod types;
-pub(crate) mod cipher_risk_client;
+mod cipher_risk_client;
 mod hibp;
+mod password_strength;
+mod types;
 
-pub use types::{
-    CipherLoginDetails, CipherRisk, CipherRiskOptions, ExposedPasswordResult, PasswordReuseMap,
-};
 pub use cipher_risk_client::{CipherRiskClient, CipherRiskError};
+pub use types::{CipherLoginDetails, CipherRisk, CipherRiskOptions, ExposedPasswordResult};
